requirements.txt
flask
flask-sqlalchemy
flask-limiter
requests
pillow
cryptography
pynput
python-dotenv
